Moved vplot into VESIcal directory and added necessary functions to __init__ to allow users to access plotting functions
